Refactor MsSqlConnection and QueryProvider
